fix: move useTheme hook to its own file to satisfy react-refresh/only-export-components lint rule
